refactor controller, update security config
