+++ src/main/java/com/c0324/casestudym5/controller/StudentController.java
@@ -1,10 +1,30 @@
 package com.c0324.casestudym5.controller;
 
+import com.c0324.casestudym5.model.Student;
+import com.c0324.casestudym5.service.StudentService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/student")
 public class StudentController {
 
+    private final StudentService studentService;
+
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    @GetMapping("/{id}")
+    public String view(@PathVariable("id") Long id, Model model, HttpSession httpSession) {
+        Student student = studentService.getStudent(id);
+        model.addAttribute("student", student);
+        model.addAttribute("pageTitle", student.getUser().getName());
+        model.addAttribute("page", httpSession.getAttribute("page"));
+        return "admin/student/student-details";
+    }
 }
