+++ src/main/java/com/c0324/casestudym5/controller/TeacherController.java
@@ -1,12 +1,35 @@
 package com.c0324.casestudym5.controller;
 
+import com.c0324.casestudym5.model.Teacher;
+import com.c0324.casestudym5.service.TeacherService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/teacher")
 public class TeacherController {
 
+    private final TeacherService teacherService;
+
+    @Autowired
+    public TeacherController(TeacherService teacherService) {
+        this.teacherService = teacherService;
+    }
+
+    @GetMapping("/detail/{id}")
+    public String getTeacher(@PathVariable Long id, Model model) {
+        Optional<Teacher> teacher = teacherService.getTeacherById(id);
+        if (teacher.isPresent()) {
+            model.addAttribute("teacher", teacher.get());
+            return "admin/teacher/teacher-details";
+        } else {
+            return "common/404";
+        }
+    }
 
 }
