+++ src/main/resources/templates/admin/teacher/teacher-list.html
@@ -56,10 +56,10 @@
                     <a th:href="@{/teacher/detail/{id}(id=${teacher.id})}" class="btn btn-info">
                         <i class="fas fa-eye"></i>
                     </a>
-                    <a th:href="@{/teacher/update/{id}(id=${teacher.id})}" class="btn btn-warning">
+                    <a th:href="@{admin/teacher/update/{id}(id=${teacher.id})}" class="btn btn-warning">
                         <i class="fas fa-edit"></i>
                     </a>
-                    <a th:href="@{/teacher/delete/{id}(id=${teacher.id})}" class="btn btn-danger">
+                    <a th:href="@{admin/teacher/delete/{id}(id=${teacher.id})}" class="btn btn-danger">
                         <i class="fas fa-trash"></i>
                     </a>
                 </td>
@@ -70,7 +70,7 @@
         <nav aria-label="Page navigation">
             <ul class="pagination justify-content-center">
                 <li class="page-item" th:each="i : ${#numbers.sequence(0, totalPages - 1)}">
-                    <a class="page-link" th:href="@{/teacher(page=${i}, size=${5}, searchQuery=${searchQuery})}" th:text="${i + 1}"></a>
+                    <a class="page-link" th:href="@{/admin/teacher(page=${i}, size=${5}, searchQuery=${searchQuery})}" th:text="${i + 1}"></a>
                 </li>
             </ul>
         </nav>
